# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Organizations API",
        "description": "API для работы с организациями, зданиями и видами деятельности",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/organizations": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Получить список всех организаций",
                "operationId": "a6f3e52e9b7995087fc40990424c2977",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Количество записей на странице",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Organization"
                                            }
                                        },
                                        "pagination": {
                                            "$ref": "#/components/schemas/Pagination"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный доступ"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/organizations/{organization}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Получить информацию об организации по ID",
                "operationId": "effcac3de5c029c2372022990ad7a48a",
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "description": "ID организации",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Organization"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Организация не найдена"
                    },
                    "401": {
                        "description": "Неавторизованный доступ"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/organizations/building/{building}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Получить список организаций в конкретном здании",
                "operationId": "d2a0668a7ebdf28d693a12222cc9b5d0",
                "parameters": [
                    {
                        "name": "building",
                        "in": "path",
                        "description": "ID здания",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Organization"
                                            }
                                        },
                                        "building": {
                                            "$ref": "#/components/schemas/Building"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Здание не найдено"
                    },
                    "401": {
                        "description": "Неавторизованный доступ"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/organizations/activity/{activity}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Получить список организаций по виду деятельности",
                "operationId": "1935eb0f0c3847cd6468b284e128d6cb",
                "parameters": [
                    {
                        "name": "activity",
                        "in": "path",
                        "description": "ID вида деятельности",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Organization"
                                            }
                                        },
                                        "activity": {
                                            "$ref": "#/components/schemas/Activity"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Вид деятельности не найден"
                    },
                    "401": {
                        "description": "Неавторизованный доступ"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/organizations/search/name": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Поиск организаций по названию",
                "operationId": "93b4e6ec0d1bfe63ef8aa56a800698f7",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Название организации для поиска",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "minLength": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Organization"
                                            }
                                        },
                                        "search_term": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    },
                    "401": {
                        "description": "Неавторизованный доступ"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/organizations/search/activity/{activity}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Поиск организаций по виду деятельности (включая дочерние)",
                "operationId": "67f9fd9e7c9df258c6d5d2df9e81636e",
                "parameters": [
                    {
                        "name": "activity",
                        "in": "path",
                        "description": "ID вида деятельности",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Organization"
                                            }
                                        },
                                        "activity": {
                                            "$ref": "#/components/schemas/Activity"
                                        },
                                        "included_activities": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Activity"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Вид деятельности не найден"
                    },
                    "401": {
                        "description": "Неавторизованный доступ"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/organizations/nearby": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Поиск организаций в радиусе или прямоугольной области",
                "operationId": "f48b9ede7a0b474a104c338c5bd06a4d",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "query",
                        "description": "Широта центра поиска",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "maximum": 90,
                            "minimum": -90
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "description": "Долгота центра поиска",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "maximum": 180,
                            "minimum": -180
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "description": "Радиус поиска в километрах",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "minimum": 0.1
                        }
                    },
                    {
                        "name": "rect_width",
                        "in": "query",
                        "description": "Ширина прямоугольной области в градусах",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "minimum": 0.1
                        }
                    },
                    {
                        "name": "rect_height",
                        "in": "query",
                        "description": "Высота прямоугольной области в градусах",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "minimum": 0.1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Organization"
                                            }
                                        },
                                        "buildings": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Building"
                                            }
                                        },
                                        "search_center": {
                                            "$ref": "#/components/schemas/Coordinates"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    },
                    "401": {
                        "description": "Неавторизованный доступ"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Building": {
                "title": "Building",
                "description": "Модель здания",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "address": {
                        "type": "string",
                        "example": "г. Москва, ул. Ленина 1, офис 3"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 55.7558
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": 37.6176
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "organizations": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Organization"
                        }
                    }
                },
                "type": "object"
            },
            "Organization": {
                "title": "Organization",
                "description": "Модель организации",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "ООО Рога и Копыта"
                    },
                    "building_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "building": {
                        "$ref": "#/components/schemas/Building"
                    },
                    "activities": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Activity"
                        }
                    },
                    "phones": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrganizationPhone"
                        }
                    }
                },
                "type": "object"
            },
            "OrganizationPhone": {
                "title": "OrganizationPhone",
                "description": "Модель телефона организации",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "organization_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "phone": {
                        "type": "string",
                        "example": "+7-123-456-78-90"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Pagination": {
                "title": "Pagination",
                "description": "Схема пагинации",
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "total": {
                        "type": "integer",
                        "example": 150
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 55.7558
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": 37.6176
                    }
                },
                "type": "object"
            },
            "Coordinates": {
                "title": "Coordinates",
                "description": "Географические координаты"
            }
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "X-API-Key",
                "in": "header"
            }
        }
    },
...
}